Initial support for carbonapi_proto_v3 (new protocol in carbonapi 0.12+)
This also fixes multiple target support for pickle
