+++ carbonserver/carbonserver.go
@@ -17,11 +17,8 @@
 package carbonserver
 
 import (
-	"bytes"
 	"context"
 	"encoding/binary"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"math"
 	"net"
@@ -31,7 +28,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -42,14 +38,12 @@ import (
 
 	"github.com/NYTimes/gziphandler"
 	"github.com/dgryski/go-expirecache"
-	trigram "github.com/dgryski/go-trigram"
+	"github.com/dgryski/go-trigram"
 	"github.com/dgryski/httputil"
-	whisper "github.com/go-graphite/go-whisper"
+	protov3 "github.com/go-graphite/protocol/carbonapi_v3_pb"
 	"github.com/lomik/go-carbon/helper"
-	pb "github.com/lomik/go-carbon/helper/carbonzipperpb"
 	"github.com/lomik/go-carbon/helper/stat"
 	"github.com/lomik/go-carbon/points"
-	pickle "github.com/lomik/og-rek"
 	"github.com/lomik/zapwriter"
 	"github.com/syndtr/goleveldb/leveldb"
 	"github.com/syndtr/goleveldb/leveldb/filter"
@@ -104,15 +98,6 @@ type QueryItem struct {
 	Data          atomic.Value
 	Flags         uint64 // DataIsAvailable or QueryIsPending
 	QueryFinished chan struct{}
-}
-
-type fetchResponse struct {
-	data           []byte
-	contentType    string
-	metricsFetched int
-	valuesFetched  int
-	memoryUsed     int
-	metrics        []string
 }
 
 var TraceHeaders = map[string]string{
@@ -263,7 +248,7 @@ type CarbonserverListener struct {
 }
 
 type metricDetailsFlat struct {
-	*pb.MetricDetails
+	*protov3.MetricDetails
 	Name string
 }
 
@@ -276,7 +261,7 @@ type jsonMetricDetailsResponse struct {
 type fileIndex struct {
 	idx     trigram.Index
 	files   []string
-	details map[string]*pb.MetricDetails
+	details map[string]*protov3.MetricDetails
 
 	accessTimes map[string]int64
 	freeSpace   uint64
@@ -378,7 +363,7 @@ func (listener *CarbonserverListener) UpdateMetricsAccessTimes(metrics map[strin
 		if _, ok := idx.details[m]; ok {
 			idx.details[m].RdTime = t
 		} else {
-			idx.details[m] = &pb.MetricDetails{RdTime: t}
+			idx.details[m] = &protov3.MetricDetails{RdTime: t}
 		}
 		idx.accessTimes[m] = t
 
@@ -435,7 +420,7 @@ func (listener *CarbonserverListener) updateFileList(dir string) {
 	t0 := time.Now()
 
 	var files []string
-	details := make(map[string]*pb.MetricDetails)
+	details := make(map[string]*protov3.MetricDetails)
 
 	metricsKnown := uint64(0)
 	err := filepath.Walk(dir, func(p string, info os.FileInfo, err error) error {
@@ -452,7 +437,7 @@ func (listener *CarbonserverListener) updateFileList(dir string) {
 			if hasSuffix {
 				metricsKnown++
 				trimmedName = strings.Replace(trimmedName[1:len(trimmedName)-4], "/", ".", -1)
-				details[trimmedName] = &pb.MetricDetails{
+				details[trimmedName] = &protov3.MetricDetails{
 					Size_:   i.RealSize,
 					ModTime: i.MTime,
 					ATime:   i.ATime,
@@ -532,7 +517,7 @@ func (listener *CarbonserverListener) updateFileList(dir string) {
 		zap.Duration("indexing_runtime", indexingRuntime),
 		zap.Duration("rdtime_update_runtime", rdTimeUpdateRuntime),
 		zap.Duration("total_runtime", time.Since(t0)),
-		zap.Int("files", len(files)),
+		zap.Int("Files", len(files)),
 		zap.Int("index_size", indexSize),
 		zap.Int("pruned_trigrams", pruned),
 	)
@@ -682,906 +667,6 @@ func (listener *CarbonserverListener) expandGlobs(query string) ([]string, []boo
 	}
 
 	return files, leafs, nil
-}
-
-var errMaxGlobsExhausted = fmt.Errorf("maxGlobs in request exhausted, kindly refusing to perform the request")
-var errMetricsListEmpty = fmt.Errorf("File index is empty or disabled")
-
-func (listener *CarbonserverListener) getMetricsList() ([]string, error) {
-	fidx := listener.CurrentFileIndex()
-	var metrics []string
-
-	if fidx == nil {
-		return nil, errMetricsListEmpty
-	}
-
-	for _, p := range fidx.files {
-		if !strings.HasSuffix(p, ".wsp") {
-			continue
-		}
-		p = p[1 : len(p)-4]
-		metrics = append(metrics, strings.Replace(p, "/", ".", -1))
-	}
-	return metrics, nil
-}
-
-func (listener *CarbonserverListener) detailsHandler(wr http.ResponseWriter, req *http.Request) {
-	// URL: /metrics/details/?format=json
-	t0 := time.Now()
-	ctx := req.Context()
-
-	atomic.AddUint64(&listener.metrics.DetailsRequests, 1)
-
-	req.ParseForm()
-	format := req.FormValue("format")
-
-	accessLogger := TraceContextToZap(ctx, listener.accessLogger.With(
-		zap.String("handler", "details"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.String("format", format),
-	))
-
-	if format != "json" && format != "protobuf" && format != "protobuf3" {
-		atomic.AddUint64(&listener.metrics.DetailsErrors, 1)
-		accessLogger.Error("details failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "unsupported format"),
-		)
-		http.Error(wr, "Bad request (unsupported format)",
-			http.StatusBadRequest)
-		return
-	}
-
-	var err error
-
-	fidx := listener.CurrentFileIndex()
-	if fidx == nil {
-		atomic.AddUint64(&listener.metrics.DetailsErrors, 1)
-		accessLogger.Error("details failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "can't fetch metrics list"),
-			zap.Error(errMetricsListEmpty),
-		)
-		http.Error(wr, fmt.Sprintf("Can't fetch metrics details: %s", err), http.StatusInternalServerError)
-		return
-	}
-
-	var b []byte
-
-	switch format {
-	case "json":
-		response := jsonMetricDetailsResponse{
-			FreeSpace:  fidx.freeSpace,
-			TotalSpace: fidx.totalSpace,
-		}
-		listener.fileIdxMutex.Lock()
-		for m, v := range fidx.details {
-			response.Metrics = append(response.Metrics, metricDetailsFlat{
-				Name:          m,
-				MetricDetails: v,
-			})
-		}
-		b, err = json.Marshal(response)
-		listener.fileIdxMutex.Unlock()
-	case "protobuf", "protobuf3":
-		listener.fileIdxMutex.Lock()
-		response := &pb.MetricDetailsResponse{
-			Metrics:    fidx.details,
-			FreeSpace:  fidx.freeSpace,
-			TotalSpace: fidx.totalSpace,
-		}
-		b, err = response.Marshal()
-		listener.fileIdxMutex.Unlock()
-	}
-
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.ListErrors, 1)
-		accessLogger.Error("details failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "response encode failed"),
-			zap.Error(err),
-		)
-		http.Error(wr, fmt.Sprintf("An internal error has occured: %s", err), http.StatusInternalServerError)
-		return
-	}
-	wr.Write(b)
-
-	accessLogger.Info("details served",
-		zap.Duration("runtime_seconds", time.Since(t0)),
-	)
-	return
-
-}
-
-func (listener *CarbonserverListener) listHandler(wr http.ResponseWriter, req *http.Request) {
-	// URL: /metrics/list/?format=json
-	t0 := time.Now()
-	ctx := req.Context()
-
-	atomic.AddUint64(&listener.metrics.ListRequests, 1)
-
-	req.ParseForm()
-	format := req.FormValue("format")
-
-	accessLogger := TraceContextToZap(ctx, listener.accessLogger.With(
-		zap.String("handler", "list"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.String("format", format),
-	))
-
-	if format != "json" && format != "protobuf" && format != "protobuf3" {
-		atomic.AddUint64(&listener.metrics.ListErrors, 1)
-		accessLogger.Error("list failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "unsupported format"),
-		)
-		http.Error(wr, "Bad request (unsupported format)",
-			http.StatusBadRequest)
-		return
-	}
-
-	var err error
-
-	metrics, err := listener.getMetricsList()
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.ListErrors, 1)
-		accessLogger.Error("list failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "can't fetch metrics list"),
-			zap.Error(err),
-		)
-		http.Error(wr, fmt.Sprintf("Can't fetch metrics list: %s", err), http.StatusInternalServerError)
-		return
-	}
-
-	var b []byte
-	response := &pb.ListMetricsResponse{Metrics: metrics}
-	switch format {
-	case "json":
-		b, err = json.Marshal(response)
-	case "protobuf":
-		b, err = response.Marshal()
-	case "protobuf3":
-		b, err = response.Marshal()
-	}
-
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.ListErrors, 1)
-		accessLogger.Error("list failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "response encode failed"),
-			zap.Error(err),
-		)
-		http.Error(wr, fmt.Sprintf("An internal error has occured: %s", err), http.StatusInternalServerError)
-		return
-	}
-	wr.Write(b)
-
-	accessLogger.Info("list served",
-		zap.Duration("runtime_seconds", time.Since(t0)),
-	)
-	return
-
-}
-
-type findResponse struct {
-	data        []byte
-	contentType string
-	files       int
-}
-
-func (listener *CarbonserverListener) findHandler(wr http.ResponseWriter, req *http.Request) {
-	// URL: /metrics/find/?local=1&format=pickle&query=the.metric.path.with.glob
-
-	t0 := time.Now()
-	ctx := req.Context()
-
-	atomic.AddUint64(&listener.metrics.FindRequests, 1)
-
-	req.ParseForm()
-	format := req.FormValue("format")
-	query := req.FormValue("query")
-
-	var response *findResponse
-
-	logger := TraceContextToZap(ctx, listener.logger.With(
-		zap.String("handler", "find"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.String("query", query),
-		zap.String("format", format),
-	))
-
-	accessLogger := TraceContextToZap(ctx, listener.accessLogger.With(
-		zap.String("handler", "find"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.String("query", query),
-		zap.String("format", format),
-	))
-
-	if format != "json" && format != "pickle" && format != "protobuf" && format != "protobuf3" {
-		atomic.AddUint64(&listener.metrics.FindErrors, 1)
-		accessLogger.Error("find failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "unsupported format"),
-			zap.Int("http_code", http.StatusBadRequest),
-		)
-		http.Error(wr, "Bad request (unsupported format)",
-			http.StatusBadRequest)
-		return
-	}
-
-	if query == "" {
-		atomic.AddUint64(&listener.metrics.FindErrors, 1)
-		accessLogger.Error("find failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "empty query"),
-			zap.Int("http_code", http.StatusBadRequest),
-		)
-		http.Error(wr, "Bad request (no query)", http.StatusBadRequest)
-		return
-	}
-
-	var err error
-	fromCache := false
-	if listener.findCacheEnabled {
-		key := query + "&" + format
-		size := uint64(100 * 1024 * 1024)
-		item := listener.findCache.getQueryItem(key, size, 300)
-		res, ok := item.FetchOrLock()
-		if !ok {
-			logger.Debug("find cache miss")
-			atomic.AddUint64(&listener.metrics.FindCacheMiss, 1)
-			response, err = listener.findMetrics(logger, t0, format, query)
-			if err != nil {
-				item.StoreAbort()
-			} else {
-				item.StoreAndUnlock(response)
-			}
-		} else if res != nil {
-			logger.Debug("query cache hit")
-			atomic.AddUint64(&listener.metrics.FindCacheHit, 1)
-			response = res.(*findResponse)
-			fromCache = true
-		}
-	} else {
-		response, err = listener.findMetrics(logger, t0, format, query)
-	}
-
-	if response == nil {
-		accessLogger.Error("find failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "internal error while processing request"),
-			zap.Error(err),
-			zap.Int("http_code", http.StatusInternalServerError),
-		)
-		http.Error(wr, fmt.Sprintf("Internal error while processing request (%v)", err),
-			http.StatusInternalServerError)
-		return
-	}
-
-	wr.Header().Set("Content-Type", response.contentType)
-	wr.Write(response.data)
-
-	if response.files == 0 {
-		// to get an idea how often we search for nothing
-		atomic.AddUint64(&listener.metrics.FindZero, 1)
-	}
-
-	accessLogger.Info("find success",
-		zap.Duration("runtime_seconds", time.Since(t0)),
-		zap.Int("files", response.files),
-		zap.Bool("find_cache_enabled", listener.findCacheEnabled),
-		zap.Bool("from_cache", fromCache),
-		zap.Int("http_code", http.StatusOK),
-	)
-	return
-}
-
-var findMetricsNoMetricsError = fmt.Errorf("no metrics available for this query")
-
-func (listener *CarbonserverListener) findMetrics(logger *zap.Logger, t0 time.Time, format, name string) (*findResponse, error) {
-	var result findResponse
-	files, leafs, err := listener.expandGlobs(name)
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.FindErrors, 1)
-
-		logger.Error("find failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "can't expand globs"),
-			zap.Error(err),
-		)
-		return nil, err
-	}
-
-	metricsCount := uint64(0)
-	for i := range files {
-		if leafs[i] {
-			metricsCount++
-		}
-	}
-	result.files = len(files)
-	atomic.AddUint64(&listener.metrics.MetricsFound, metricsCount)
-	logger.Debug("expandGlobs result",
-		zap.String("action", "expandGlobs"),
-		zap.String("metric", name),
-		zap.Uint64("metrics_count", metricsCount),
-	)
-
-	if format == "json" || format == "protobuf" || format == "protobuf3" {
-		var err error
-		response := pb.GlobResponse{
-			Name:    name,
-			Matches: make([]*pb.GlobMatch, 0),
-		}
-
-		for i, p := range files {
-			response.Matches = append(response.Matches, &pb.GlobMatch{Path: p, IsLeaf: leafs[i]})
-		}
-
-		switch format {
-		case "json":
-			result.contentType = "application/json"
-			result.data, err = json.Marshal(response)
-		case "protobuf3", "protobuf":
-			result.contentType = "application/protobuf"
-			result.data, err = response.Marshal()
-		}
-
-		if err != nil {
-			atomic.AddUint64(&listener.metrics.FindErrors, 1)
-
-			logger.Error("find failed",
-				zap.Duration("runtime_seconds", time.Since(t0)),
-				zap.String("reason", "response encode failed"),
-				zap.Error(err),
-			)
-			return nil, err
-		} else {
-			if len(response.Matches) == 0 {
-				err = findMetricsNoMetricsError
-			}
-		}
-		return &result, err
-	} else if format == "pickle" {
-		// [{'metric_path': 'metric', 'intervals': [(x,y)], 'isLeaf': True},]
-		var metrics []map[string]interface{}
-		var m map[string]interface{}
-
-		intervals := &IntervalSet{Start: 0, End: int32(time.Now().Unix()) + 60}
-
-		for i, p := range files {
-			m = make(map[string]interface{})
-			if listener.graphiteweb10 {
-				// graphite master
-				m["path"] = p
-				m["is_leaf"] = leafs[i]
-				m["intervals"] = intervals
-			} else {
-				// graphite 0.9.x
-				m["metric_path"] = p
-				m["isLeaf"] = leafs[i]
-			}
-
-			metrics = append(metrics, m)
-		}
-		var buf bytes.Buffer
-		pEnc := pickle.NewEncoder(&buf)
-		pEnc.Encode(metrics)
-		return &findResponse{buf.Bytes(), "application/pickle", len(files)}, nil
-	}
-	return nil, nil
-}
-
-func (listener *CarbonserverListener) renderHandler(wr http.ResponseWriter, req *http.Request) {
-	// URL: /render/?target=the.metric.name&format=pickle&from=1396008021&until=1396022421
-	t0 := time.Now()
-	ctx := req.Context()
-
-	atomic.AddUint64(&listener.metrics.RenderRequests, 1)
-
-	req.ParseForm()
-	targets := req.Form["target"]
-	format := req.FormValue("format")
-	from := req.FormValue("from")
-	until := req.FormValue("until")
-
-	logger := TraceContextToZap(ctx, listener.accessLogger.With(
-		zap.String("handler", "render"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.Strings("targets", targets),
-		zap.String("from", from),
-		zap.String("until", until),
-		zap.String("format", format),
-	))
-
-	accessLogger := TraceContextToZap(ctx, listener.accessLogger.With(
-		zap.String("handler", "render"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.Strings("targets", targets),
-		zap.String("from", from),
-		zap.String("until", until),
-		zap.String("format", format),
-	))
-
-	// Make sure we log which metric caused a panic()
-	defer func() {
-		if r := recover(); r != nil {
-			logger.Error("panic recovered",
-				zap.Stack("stack"),
-				zap.Any("error", r),
-			)
-			accessLogger.Error("fetch failed",
-				zap.Duration("runtime_seconds", time.Since(t0)),
-				zap.String("reason", "panic during serving the request"),
-				zap.Stack("stack"),
-				zap.Any("error", r),
-				zap.Int("http_code", http.StatusInternalServerError),
-			)
-			http.Error(wr, fmt.Sprintf("Panic occured, see logs for more information"),
-				http.StatusInternalServerError)
-		}
-	}()
-
-	if format != "json" && format != "pickle" && format != "protobuf" && format != "protobuf3" {
-		atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-		accessLogger.Error("fetch failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "unsupported format"),
-			zap.Int("http_code", http.StatusBadRequest),
-		)
-		http.Error(wr, "Bad request (unsupported format)",
-			http.StatusBadRequest)
-		return
-	}
-
-	response, fromCache, err := listener.fetchWithCache(logger, format, targets, from, until)
-
-	wr.Header().Set("Content-Type", response.contentType)
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-		accessLogger.Error("fetch failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "failed to read data"),
-			zap.Int("http_code", http.StatusBadRequest),
-			zap.Error(err),
-		)
-		http.Error(wr, fmt.Sprintf("Bad request (%s)", err),
-			http.StatusBadRequest)
-		return
-	}
-
-	listener.UpdateMetricsAccessTimesByRequest(response.metrics)
-
-	wr.Write(response.data)
-
-	atomic.AddUint64(&listener.metrics.FetchSize, uint64(response.memoryUsed))
-	logger.Info("fetch served",
-		zap.Duration("runtime_seconds", time.Since(t0)),
-		zap.Bool("query_cache_enabled", listener.queryCacheEnabled),
-		zap.Bool("from_cache", fromCache),
-		zap.Int("metrics_fetched", response.metricsFetched),
-		zap.Int("values_fetched", response.valuesFetched),
-		zap.Int("memory_used_bytes", response.memoryUsed),
-		zap.Int("http_code", http.StatusOK),
-	)
-
-}
-
-func (listener *CarbonserverListener) fetchWithCache(logger *zap.Logger, format string, targets []string, from, until string) (fetchResponse, bool, error) {
-	logger = logger.With(
-		zap.String("function", "fetchWithCache"),
-	)
-	fromCache := false
-	var badTime bool
-
-	i, err := strconv.Atoi(from)
-	if err != nil {
-		logger.Debug("invalid fromTime", zap.Error(err))
-		badTime = true
-	}
-	fromTime := int32(i)
-	i, err = strconv.Atoi(until)
-	if err != nil {
-		logger.Debug("invalid untilTime", zap.Error(err))
-		badTime = true
-	}
-	untilTime := int32(i)
-
-	if badTime {
-		atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-		err = errors.New("Bad request (invalid from/until time)")
-		return fetchResponse{nil, "application/text", 0, 0, 0, nil}, fromCache, err
-	}
-
-	var response fetchResponse
-	if listener.queryCacheEnabled {
-		key := strings.Join(targets, "&") + "&" + format + "&" + from + "&" + until
-		size := uint64(100 * 1024 * 1024)
-		renderRequests := atomic.LoadUint64(&listener.metrics.RenderRequests)
-		fetchSize := atomic.LoadUint64(&listener.metrics.FetchSize)
-		if renderRequests > 0 {
-			size = fetchSize / renderRequests
-		}
-		item := listener.queryCache.getQueryItem(key, size, 60)
-		res, ok := item.FetchOrLock()
-		if !ok {
-			logger.Debug("query cache miss")
-			atomic.AddUint64(&listener.metrics.QueryCacheMiss, 1)
-			response, err = listener.prepareData(format, targets, fromTime, untilTime)
-			if err != nil {
-				item.StoreAbort()
-			} else {
-				item.StoreAndUnlock(response)
-			}
-		} else {
-			logger.Debug("query cache hit")
-			atomic.AddUint64(&listener.metrics.QueryCacheHit, 1)
-			response = res.(fetchResponse)
-			fromCache = true
-		}
-	} else {
-		response, err = listener.prepareData(format, targets, fromTime, untilTime)
-	}
-	return response, fromCache, err
-}
-
-func (listener *CarbonserverListener) prepareData(format string, targets []string, fromTime, untilTime int32) (fetchResponse, error) {
-	contentType := "application/text"
-	var b []byte
-	var err error
-	var metricsFetched int
-	var memoryUsed int
-	var valuesFetched int
-
-	var multi pb.MultiFetchResponse
-	for _, metric := range targets {
-		listener.logger.Debug("fetching data...")
-		files, leafs, err := listener.expandGlobs(metric)
-		if err != nil {
-			listener.logger.Debug("expand globs returned an error",
-				zap.Error(err),
-			)
-			continue
-		}
-
-		metricsCount := 0
-		for i := range files {
-			if leafs[i] {
-				metricsCount++
-			}
-		}
-		listener.logger.Debug("expandGlobs result",
-			zap.String("handler", "render"),
-			zap.String("action", "expandGlobs"),
-			zap.String("metric", metric),
-			zap.Int("metrics_count", metricsCount),
-			zap.Int32("from", fromTime),
-			zap.Int32("until", untilTime),
-		)
-
-		res, err := listener.fetchDataPB(metric, files, leafs, fromTime, untilTime)
-		if err != nil {
-			atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-			listener.logger.Error("error while fetching the data",
-				zap.Error(err),
-			)
-			continue
-		}
-
-		multi.Metrics = append(multi.Metrics, res.Metrics...)
-	}
-	if len(multi.Metrics) == 0 {
-		return fetchResponse{nil, contentType, 0, 0, 0, nil}, err
-	}
-
-	var metrics []string
-	metricsFetched = len(multi.Metrics)
-	for i := range multi.Metrics {
-		metrics = append(metrics, multi.Metrics[i].Name)
-		memoryUsed += multi.Metrics[i].Size()
-		valuesFetched += len(multi.Metrics[i].Values)
-	}
-
-	switch format {
-	case "json":
-		contentType = "application/json"
-		b, err = json.Marshal(multi)
-	case "protobuf3", "protobuf":
-		contentType = "application/protobuf"
-		b, err = multi.Marshal()
-	case "pickle":
-		// transform protobuf data into what pickle expects
-		//[{'start': 1396271100, 'step': 60, 'name': 'metric',
-		//'values': [9.0, 19.0, None], 'end': 1396273140}
-
-		var response []map[string]interface{}
-
-		for _, metric := range multi.GetMetrics() {
-
-			var m map[string]interface{}
-
-			m = make(map[string]interface{})
-			m["start"] = metric.StartTime
-			m["step"] = metric.StepTime
-			m["end"] = metric.StopTime
-			m["name"] = metric.Name
-
-			mv := make([]interface{}, len(metric.Values))
-			for i, p := range metric.Values {
-				if metric.IsAbsent[i] {
-					mv[i] = nil
-				} else {
-					mv[i] = p
-				}
-			}
-
-			m["values"] = mv
-			response = append(response, m)
-		}
-
-		contentType = "application/pickle"
-		var buf bytes.Buffer
-		pEnc := pickle.NewEncoder(&buf)
-		err = pEnc.Encode(response)
-		b = buf.Bytes()
-	}
-
-	if err != nil {
-		return fetchResponse{nil, contentType, 0, 0, 0, nil}, err
-	}
-	return fetchResponse{b, contentType, metricsFetched, valuesFetched, memoryUsed, metrics}, nil
-}
-
-func (listener *CarbonserverListener) fetchSingleMetric(metric string, fromTime, untilTime int32) (*pb.FetchResponse, error) {
-	var step int32
-
-	// We need to obtain the metadata from whisper file anyway.
-	path := listener.whisperData + "/" + strings.Replace(metric, ".", "/", -1) + ".wsp"
-	w, err := whisper.OpenWithOptions(path, &whisper.Options{
-		FLock: listener.flock,
-	})
-	if err != nil {
-		// the FE/carbonzipper often requests metrics we don't have
-		// We shouldn't really see this any more -- expandGlobs() should filter them out
-		atomic.AddUint64(&listener.metrics.NotFound, 1)
-		listener.logger.Error("open error", zap.String("path", path), zap.Error(err))
-		return nil, errors.New("Can't open metric")
-	}
-
-	logger := listener.logger.With(
-		zap.String("path", path),
-		zap.Int("fromTime", int(fromTime)),
-		zap.Int("untilTime", int(untilTime)),
-	)
-
-	retentions := w.Retentions()
-	now := int32(time.Now().Unix())
-	diff := now - fromTime
-	bestStep := int32(retentions[0].SecondsPerPoint())
-	for _, retention := range retentions {
-		if int32(retention.MaxRetention()) >= diff {
-			step = int32(retention.SecondsPerPoint())
-			break
-		}
-	}
-
-	if step == 0 {
-		maxRetention := int32(retentions[len(retentions)-1].MaxRetention())
-		if now-maxRetention > untilTime {
-			atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-			logger.Warn("can't find proper archive for the request")
-			return nil, errors.New("Can't find proper archive")
-		}
-		logger.Debug("can't find archive that contains full set of data, using the least precise one")
-		step = maxRetention
-	}
-
-	var cacheData []points.Point
-	if step != bestStep {
-		logger.Debug("cache is not supported for this query (required step != best step)",
-			zap.Int("step", int(step)),
-			zap.Int("bestStep", int(bestStep)),
-		)
-	} else {
-		// query cache
-		cacheStartTime := time.Now()
-		cacheData = listener.cacheGet(metric)
-		waitTime := uint64(time.Since(cacheStartTime).Nanoseconds())
-		atomic.AddUint64(&listener.metrics.CacheWaitTimeFetchNS, waitTime)
-	}
-
-	logger.Debug("fetching disk metric")
-	atomic.AddUint64(&listener.metrics.DiskRequests, 1)
-	diskStartTime := time.Now()
-	points, err := w.Fetch(int(fromTime), int(untilTime))
-	w.Close()
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-		logger.Warn("failed to fetch points", zap.Error(err))
-		return nil, errors.New("failed to fetch points")
-	}
-
-	// Should never happen, because we have a check for proper archive now
-	if points == nil {
-		atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-		logger.Warn("metric time range not found")
-		return nil, errors.New("time range not found")
-	}
-	atomic.AddUint64(&listener.metrics.MetricsReturned, 1)
-	values := points.Values()
-
-	fromTime = int32(points.FromTime())
-	untilTime = int32(points.UntilTime())
-	step = int32(points.Step())
-
-	waitTime := uint64(time.Since(diskStartTime).Nanoseconds())
-	atomic.AddUint64(&listener.metrics.DiskWaitTimeNS, waitTime)
-	atomic.AddUint64(&listener.metrics.PointsReturned, uint64(len(values)))
-
-	response := pb.FetchResponse{
-		Name:      metric,
-		StartTime: fromTime,
-		StopTime:  untilTime,
-		StepTime:  step,
-		Values:    make([]float64, len(values)),
-		IsAbsent:  make([]bool, len(values)),
-	}
-
-	for i, p := range values {
-		if math.IsNaN(p) {
-			response.Values[i] = 0
-			response.IsAbsent[i] = true
-		} else {
-			response.Values[i] = p
-			response.IsAbsent[i] = false
-		}
-	}
-
-	if cacheData != nil {
-		atomic.AddUint64(&listener.metrics.CacheRequestsTotal, 1)
-		cacheStartTime := time.Now()
-		pointsFetchedFromCache := 0
-		for _, item := range cacheData {
-			ts := int32(item.Timestamp) - int32(item.Timestamp)%step
-			if ts < fromTime || ts >= untilTime {
-				continue
-			}
-			pointsFetchedFromCache++
-			index := (ts - fromTime) / step
-			response.Values[index] = item.Value
-			response.IsAbsent[index] = false
-		}
-		waitTime := uint64(time.Since(cacheStartTime).Nanoseconds())
-		atomic.AddUint64(&listener.metrics.CacheWorkTimeNS, waitTime)
-		if pointsFetchedFromCache > 0 {
-			atomic.AddUint64(&listener.metrics.CacheHit, 1)
-		} else {
-			atomic.AddUint64(&listener.metrics.CacheMiss, 1)
-		}
-	}
-	return &response, nil
-}
-
-func (listener *CarbonserverListener) fetchDataPB(metric string, files []string, leafs []bool, fromTime, untilTime int32) (*pb.MultiFetchResponse, error) {
-	var multi pb.MultiFetchResponse
-	for i, metric := range files {
-		if !leafs[i] {
-			listener.logger.Debug("skipping directory", zap.String("metric", metric))
-			// can't fetch a directory
-			continue
-		}
-		response, err := listener.fetchSingleMetric(metric, fromTime, untilTime)
-		if err == nil {
-			multi.Metrics = append(multi.Metrics, response)
-		}
-	}
-	return &multi, nil
-}
-
-func (listener *CarbonserverListener) infoHandler(wr http.ResponseWriter, req *http.Request) {
-	// URL: /info/?target=the.metric.name&format=json
-	t0 := time.Now()
-	ctx := req.Context()
-
-	atomic.AddUint64(&listener.metrics.InfoRequests, 1)
-	req.ParseForm()
-	metric := req.FormValue("target")
-	format := req.FormValue("format")
-
-	accessLogger := TraceContextToZap(ctx, listener.accessLogger.With(
-		zap.String("handler", "info"),
-		zap.String("url", req.URL.RequestURI()),
-		zap.String("peer", req.RemoteAddr),
-		zap.String("target", metric),
-		zap.String("format", format),
-	))
-
-	if format == "" {
-		format = "json"
-	}
-
-	if format != "json" && format != "protobuf" && format != "protobuf3" {
-		atomic.AddUint64(&listener.metrics.InfoErrors, 1)
-		accessLogger.Error("info failed",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "unsupported format"),
-			zap.Int("http_code", http.StatusBadRequest),
-		)
-		http.Error(wr, "Bad request (unsupported format)",
-			http.StatusBadRequest)
-		return
-	}
-
-	path := listener.whisperData + "/" + strings.Replace(metric, ".", "/", -1) + ".wsp"
-	w, err := whisper.Open(path)
-
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.NotFound, 1)
-		accessLogger.Error("info served",
-			zap.Duration("runtime_seconds", time.Since(t0)),
-			zap.String("reason", "metric not found"),
-			zap.Int("http_code", http.StatusNotFound),
-		)
-		http.Error(wr, "Metric not found", http.StatusNotFound)
-		return
-	}
-
-	defer w.Close()
-
-	aggr := w.AggregationMethod()
-	maxr := int32(w.MaxRetention())
-	xfiles := float32(w.XFilesFactor())
-
-	var b []byte
-	rets := make([]*pb.Retention, 0, 4)
-	for _, retention := range w.Retentions() {
-		spp := int32(retention.SecondsPerPoint())
-		nop := int32(retention.NumberOfPoints())
-		rets = append(rets, &pb.Retention{
-			SecondsPerPoint: spp,
-			NumberOfPoints:  nop,
-		})
-	}
-
-	response := pb.InfoResponse{
-		Name:              metric,
-		AggregationMethod: aggr,
-		MaxRetention:      maxr,
-		XFilesFactor:      xfiles,
-		Retentions:        rets,
-	}
-
-	switch format {
-	case "json":
-		b, err = json.Marshal(response)
-	case "protobuf3", "protobuf":
-		b, err = response.Marshal()
-	}
-
-	if err != nil {
-		atomic.AddUint64(&listener.metrics.RenderErrors, 1)
-		accessLogger.Error("info failed",
-			zap.String("reason", "response encode failed"),
-			zap.Int("http_code", http.StatusInternalServerError),
-			zap.Error(err),
-		)
-		http.Error(wr, "Failed to encode response: "+err.Error(),
-			http.StatusInternalServerError)
-		return
-	}
-	wr.Write(b)
-
-	accessLogger.Info("info served",
-		zap.Duration("runtime_seconds", time.Since(t0)),
-		zap.Int("http_code", http.StatusOK),
-	)
-	return
 }
 
 func (listener *CarbonserverListener) Stat(send helper.StatCallback) {
